// frontend/src/app/models/orden-trabajo.model.ts
export interface OrdenTrabajo {
    OT_num?: number;
    codigo_unico?: string;  
    operario_username: string;
    fecha: string;  
    observacion: string;
    edificio_nombre: string;
    tarea_descripcion: string;
    sector_nombre: string;
    piso_nombre: string;
    ubicacion_descripcion: string;
    tipo_activo: string;
    tareas: string;
  }
  